# apps/api/openapi/mimorii.openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/auth/me": {
      "get": {
        "operationId": "current",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Authentication"
        ]
      }
    },
...
    "/api/teams/{teamId}/checks/{id}/history": {
      "get": {
        "operationId": "history",
        "parameters": [
          {
            "name": "teamId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "from",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Checks"
        ]
      }
    },
...
    "/api/teams/{teamId}/incidents": {
      "get": {
        "operationId": "list",
        "parameters": [
          {
            "name": "teamId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Incidents"
        ]
      },
      "post": {
        "operationId": "create",
        "parameters": [
          {
            "name": "teamId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateIncidentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Incidents"
        ]
      }
    },
    "/api/teams/{teamId}/incidents/{id}": {
      "get": {
        "operationId": "get",
        "parameters": [
          {
            "name": "teamId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Incidents"
        ]
      },
      "patch": {
        "operationId": "update",
        "parameters": [
          {
            "name": "teamId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateIncidentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Incidents"
        ]
      }
    },
    "/api/teams/{teamId}/incidents/{id}/updates": {
      "post": {
        "operationId": "addUpdate",
        "parameters": [
          {
            "name": "teamId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddIncidentUpdateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Incidents"
        ]
      }
    },
...
    "/api/teams/{teamId}/heartbeats": {
      "get": {
        "operationId": "list",
        "parameters": [
          {
            "name": "teamId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "resourceId",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Heartbeats"
        ]
      },
      "post": {
        "operationId": "create",
        "parameters": [
          {
            "name": "teamId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateHeartbeatMonitorDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Heartbeats"
        ]
      }
    },
    "/api/teams/{teamId}/heartbeats/{id}": {
      "get": {
        "operationId": "get",
        "parameters": [
          {
            "name": "teamId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Heartbeats"
        ]
      },
      "patch": {
        "operationId": "update",
        "parameters": [
          {
            "name": "teamId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateHeartbeatMonitorDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Heartbeats"
        ]
      },
      "delete": {
        "operationId": "remove",
        "parameters": [
          {
            "name": "teamId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Heartbeats"
        ]
      }
    },
    "/api/teams/{teamId}/heartbeats/{id}/rotate-token": {
      "post": {
        "operationId": "rotateToken",
        "parameters": [
          {
            "name": "teamId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Heartbeats"
        ]
      }
    },
    "/api/teams/{teamId}/heartbeats/{id}/history": {
      "get": {
        "operationId": "history",
        "parameters": [
          {
            "name": "teamId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 1000,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Heartbeats"
        ]
      }
    },
    "/api/heartbeats/{token}": {
      "post": {
        "operationId": "succeeded",
        "parameters": [
          {
            "name": "token",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HeartbeatSignalDto"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": ""
          }
        },
        "tags": [
          "Heartbeat ingestion"
        ]
      }
    },
    "/api/heartbeats/{token}/start": {
      "post": {
        "operationId": "started",
        "parameters": [
          {
            "name": "token",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HeartbeatSignalDto"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": ""
          }
        },
        "tags": [
          "Heartbeat ingestion"
        ]
      }
    },
    "/api/heartbeats/{token}/fail": {
      "post": {
        "operationId": "failed",
        "parameters": [
          {
            "name": "token",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HeartbeatSignalDto"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": ""
          }
        },
        "tags": [
          "Heartbeat ingestion"
        ]
      }
    },
...
    "/api/agent/heartbeat": {
      "post": {
        "operationId": "heartbeat",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentHeartbeatDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "agent-key": []
          }
        ],
        "tags": [
          "Agent transport"
        ]
      }
    },
...
  },
  "info": {
    "title": "Mimorii API",
    "description": "Self-hosted uptime, server analytics, and relay-agent API.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "User access token"
      },
      "agent-key": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Unique agent enrollment key"
      }
    },
    "schemas": {
...
      "CreateIncidentDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 160
          },
          "impact": {
            "type": "string",
            "enum": [
              "minor",
              "major",
              "critical"
            ]
          },
          "resourceIds": {
            "minItems": 1,
            "maxItems": 200,
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "message": {
            "type": "string",
            "minLength": 1,
            "maxLength": 2000
          },
          "status": {
            "type": "string",
            "enum": [
              "investigating",
              "identified",
              "monitoring"
            ]
          },
          "startedAt": {
            "type": "string"
          }
        },
        "required": [
          "title",
          "impact",
          "resourceIds",
          "message"
        ]
      },
      "UpdateIncidentDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 160
          },
          "impact": {
            "type": "string",
            "enum": [
              "minor",
              "major",
              "critical"
            ]
          },
          "resourceIds": {
            "minItems": 1,
            "maxItems": 200,
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "message": {
            "type": "string",
            "minLength": 1,
            "maxLength": 2000
          },
          "status": {
            "type": "string",
            "enum": [
              "investigating",
              "identified",
              "monitoring"
            ]
          },
          "startedAt": {
            "type": "string"
          }
        }
      },
      "AddIncidentUpdateDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "investigating",
              "identified",
              "monitoring",
              "resolved"
            ]
          },
          "message": {
            "type": "string",
            "minLength": 1,
            "maxLength": 2000
          }
        },
        "required": [
          "status",
          "message"
        ]
      },
      "CreateHeartbeatMonitorDto": {
        "type": "object",
        "properties": {
          "resourceId": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          },
          "intervalSeconds": {
            "type": "number",
            "minimum": 60,
            "maximum": 2592000,
            "default": 300
          },
          "graceSeconds": {
            "type": "number",
            "minimum": 0,
            "maximum": 86400,
            "default": 60
          },
          "maxRuntimeSeconds": {
            "type": "object",
            "minimum": 60,
            "maximum": 2592000,
            "nullable": true
          },
          "enabled": {
            "type": "boolean",
            "default": true
          }
        },
        "required": [
          "resourceId",
          "name",
          "intervalSeconds"
        ]
      },
      "UpdateHeartbeatMonitorDto": {
        "type": "object",
        "properties": {
          "resourceId": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          },
          "intervalSeconds": {
            "type": "number",
            "minimum": 60,
            "maximum": 2592000,
            "default": 300
          },
          "graceSeconds": {
            "type": "number",
            "minimum": 0,
            "maximum": 86400,
            "default": 60
          },
          "maxRuntimeSeconds": {
            "type": "object",
            "minimum": 60,
            "maximum": 2592000,
            "nullable": true
          },
          "enabled": {
            "type": "boolean",
            "default": true
          }
        }
      },
      "HeartbeatSignalDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "maxLength": 500
          },
          "durationMs": {
            "type": "number",
            "minimum": 0,
            "maximum": 2592000000
          },
          "metadata": {
            "type": "object",
            "additionalProperties": true
          }
        }
      },
...
      "DiskSnapshotDto": {
        "type": "object",
        "properties": {
          "mount": {
            "type": "string"
          },
          "usedBytes": {
            "type": "number"
          },
          "totalBytes": {
            "type": "number"
          }
        },
        "required": [
          "mount",
          "usedBytes",
          "totalBytes"
        ]
      },
      "TechnologySnapshotDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 80
          },
          "category": {
            "type": "string",
            "enum": [
              "runtime",
              "framework",
              "database",
              "proxy",
              "container",
              "protocol",
              "other"
            ]
          },
          "version": {
            "type": "object",
            "nullable": true,
            "maxLength": 80
          }
        },
        "required": [
          "name",
          "category"
        ]
      },
      "HostSnapshotDto": {
        "type": "object",
        "properties": {
          "hostname": {
            "type": "string"
          },
          "platform": {
            "type": "string"
          },
          "version": {
            "type": "string"
          },
          "uptimeSeconds": {
            "type": "number"
          },
          "cpuPercent": {
            "type": "number"
          },
          "loadAverage": {
            "type": "number"
          },
          "memoryUsedBytes": {
            "type": "number"
          },
          "memoryTotalBytes": {
            "type": "number"
          },
          "swapUsedBytes": {
            "type": "number"
          },
          "swapTotalBytes": {
            "type": "number"
          },
          "processCount": {
            "type": "number"
          },
          "networkReceivedBytes": {
            "type": "number"
          },
          "networkTransmittedBytes": {
            "type": "number"
          },
          "disks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DiskSnapshotDto"
            }
          },
          "technologies": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TechnologySnapshotDto"
            }
          },
          "observedAt": {
            "type": "string"
          }
        },
        "required": [
          "hostname",
          "platform",
          "version",
          "uptimeSeconds",
          "cpuPercent",
          "loadAverage",
          "memoryUsedBytes",
          "memoryTotalBytes",
          "swapUsedBytes",
          "swapTotalBytes",
          "processCount",
          "networkReceivedBytes",
          "networkTransmittedBytes",
          "disks",
          "technologies",
          "observedAt"
        ]
      },
      "AgentTaskResultDto": {
        "type": "object",
        "properties": {
          "taskId": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "up",
              "degraded",
              "down"
            ]
          },
          "latencyMs": {
            "type": "object",
            "nullable": true
          },
          "statusCode": {
            "type": "object",
            "nullable": true
          },
          "message": {
            "type": "object",
            "nullable": true,
            "maxLength": 500
          },
          "metrics": {
            "type": "object",
            "additionalProperties": true
          },
          "checkedAt": {
            "type": "string"
          }
        },
        "required": [
          "taskId",
          "status",
          "metrics",
          "checkedAt"
        ]
      },
      "AgentHeartbeatDto": {
        "type": "object",
        "properties": {
          "snapshots": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/HostSnapshotDto"
            }
          },
          "results": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AgentTaskResultDto"
            }
          },
          "capabilities": {
            "maxItems": 20,
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "snapshots",
          "results",
          "capabilities"
        ]
      },
...
    }
  }
}